scripts: handle image format
Looking at file extension to determine the image format does not always
work.  e.g., cloud-init images come with an .img extension but are qcow2
by default. This would result in a boot failure if we try to pass the
format as "raw" instead of qcow2.

Instead, use `qemu-img` tool that determines the image format by actually
looking into the file contents.

Signed-off-by: Vikram Narayanan <vikram186@gmail.com>
Reviewed-by: Carlos Bilbao <carlos.bilbao@amd.com>