cpu/vc: Modify the vc_terminate wrappers to return the NeverType
Some functions are just wrappers to vc_terminate(), this patch
modifies them to return the NeverType (!) as well. That allow us, for
example, to assign a match statement to a variable when the Err() or the
None branch only calls a vc_terminate wrapper. Otherwise, rustc fails to
compile the match statement.

Signed-off-by: Claudio Carvalho <cclaudio@linux.ibm.com>