vtpm: Request a vmpl0 attestation report
An attestation report can be requested by sending an specific
SNP_GUEST_REQUEST message type to the PSP firmware. This type of message
accepts two attributes: a vmpl level and a user data. If accepted by the
PSP firmware, it returns a signed attestation report to the caller. The
report contains both the vmpl level and user data provided as well as
some additional information.

This patch requests an attestation report for vmpl0 and also providing a
SHA512 hash of the TPM EKpub as user data. The idea behind that is to
use a vmpl0 attestation report to establish a unique bond between the
TPM EK and the guest. Such artifact could later be used to prove the
authenticity of the guest.

Signed-off-by: Claudio Carvalho <cclaudio@linux.ibm.com>