psp/guest_request_cmd: Initialize SNP Guest Request resources
To send a SNP Guest Request command to the PSP, we need to provide the
request and response physical addresses to the PSP; depending on the
command we also need to provide the physical address of a data buffer.
The size of these buffers are 1, 1 and 4 pages, and they need to be
shared with the hypervisor.

The attestation report is on the critical path to attest the
confidential VM. This patch pre-allocates the required buffers at boot
time to prevent us from having to handle out-of-memory issues later at
boot time or runtime.

Access to the buffers are controled through a spinlock.

Signed-off-by: Claudio Carvalho <cclaudio@linux.ibm.com>